[Move] configuraion interfaces into new file
